pyproject.toml
[tool.poetry]
name = "infobip_api_client"
version = "5.1.0"
description = "This is a Python package for Infobip API and you can use it as a dependency to add Infobip APIs to your application."
authors = ["Infobip support <support@infobip.com>"]
license = "MIT License"
readme = "README.md"
repository = "https://github.com/infobip/infobip-api-python-client.git"
keywords = ["infobip", "sms", "tfa", "2fa", "calls", "voice", "sdk", "rest", "api",  "openapi"]
include = ["infobip_api_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"

python_dateutil = ">=2.5.3"
setuptools = ">=72.1.0"
urllib3 = ">=1.25.3, < 2.1.0"
pydantic = ">=2"
typing-extensions = ">=4.7.1"

[tool.poetry.dev-dependencies]
pytest="6.2.5"
pytest-httpserver="1.0.8"
pytest-cases="3.6.8"
pytest-cov="3.0.0"
werkzeug="3.0.3"


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"

[tool.mypy]
files = [
  "infobip_api_client",
  #"test",  # auto-generated tests
  "tests", # hand-written tests
]
# TODO: enable "strict" once all these individual checks are passing
# strict = true

# List from: https://mypy.readthedocs.io/en/stable/existing_code.html#introduce-stricter-options
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

## Getting these passing should be easy
strict_equality = true
strict_concatenate = true

## Strongly recommend enabling this one as soon as you can
check_untyped_defs = true

## These shouldn't be too much additional work, but may be tricky to
## get passing if you use a lot of untyped libraries
disallow_subclassing_any = true
disallow_untyped_decorators = true
disallow_any_generics = true

### These next few are various gradations of forcing use of type annotations
#disallow_untyped_calls = true
#disallow_incomplete_defs = true
#disallow_untyped_defs = true
#
### This one isn't too hard to get passing, but return on investment is lower
#no_implicit_reexport = true
#
### This one can be tricky to get passing if you use a lot of untyped libraries
#warn_return_any = true
